refactor(commands): extract usage-reporting commands into usage.rs
