+++ src-rust/crates/commands/src/lib.rs
@@ -258,6 +258,8 @@ mod diagnostics;
 pub use diagnostics::*;
 mod providers;
 pub use providers::*;
+mod usage;
+pub use usage::*;
 mod ui_settings;
 use ui_settings::*;
 
@@ -276,14 +278,12 @@ pub struct ResumeCommand;
 pub struct StatusCommand;
 pub struct DiffCommand;
 pub struct BugCommand;
-pub struct UsageCommand;
 pub struct InitCommand;
 pub struct HooksCommand;
 pub struct ImportConfigCommand;
 pub struct ThinkingCommand;
 // New commands
 // Batch-1 new commands
-pub struct ExtraUsageCommand;
 pub struct FastCommand;
 pub struct FeedbackCommand;
 pub struct ColorSetCommand;
@@ -1018,68 +1018,6 @@ impl SlashCommand for BugCommand {
     }
 }
 
-// ---- /usage --------------------------------------------------------------
-
-#[async_trait]
-impl SlashCommand for UsageCommand {
-    fn name(&self) -> &str { "usage" }
-    fn description(&self) -> &str { "Show API usage, quotas, and rate limit status" }
-    fn help(&self) -> &str {
-        "Usage: /usage\n\n\
-         Shows current session API usage and account quota information.\n\
-         For detailed per-call breakdown, use /extra-usage.\n\
-         For cost details, use /cost."
-    }
-
-    async fn execute(&self, _args: &str, ctx: &mut CommandContext) -> CommandResult {
-        let input = ctx.cost_tracker.input_tokens();
-        let output = ctx.cost_tracker.output_tokens();
-        let cache_creation = ctx.cost_tracker.cache_creation_tokens();
-        let cache_read = ctx.cost_tracker.cache_read_tokens();
-        let total = ctx.cost_tracker.total_tokens();
-        let cost = ctx.cost_tracker.total_cost_usd();
-
-        // Try to get account tier from OAuth tokens
-        let account_info = match claurst_core::oauth::OAuthTokens::load().await {
-            Some(tokens) => {
-                let sub = tokens.subscription_type.as_deref().unwrap_or("unknown");
-                format!("Plan: {}", sub)
-            }
-            None => {
-                if ctx.config.resolve_api_key().is_some() {
-                    "Plan: API key (Console billing)".to_string()
-                } else {
-                    "Plan: not authenticated — run /login".to_string()
-                }
-            }
-        };
-
-        CommandResult::Message(format!(
-            "API Usage — Current Session\n\
-             ────────────────────────────\n\
-             {account_info}\n\
-             Model:          {model}\n\n\
-             Tokens used this session:\n\
-               Input:        {input:>10}\n\
-               Output:       {output:>10}\n\
-               Cache write:  {cache_creation:>10}\n\
-               Cache read:   {cache_read:>10}\n\
-               Total:        {total:>10}\n\n\
-             Estimated cost: ${cost:.4}\n\n\
-             Use /extra-usage for per-call breakdown.\n\
-             Use /rate-limit-options to see your plan limits.",
-            account_info = account_info,
-            model = ctx.config.effective_model(),
-            input = input,
-            output = output,
-            cache_creation = cache_creation,
-            cache_read = cache_read,
-            total = total,
-            cost = cost,
-        ))
-    }
-}
-
 // ---- /init ---------------------------------------------------------------
 
 #[async_trait]
@@ -1196,89 +1134,6 @@ impl SlashCommand for ThinkingCommand {
                 model
             ))
         }
-    }
-}
-
-// ---- /extra-usage --------------------------------------------------------
-
-#[async_trait]
-impl SlashCommand for ExtraUsageCommand {
-    fn name(&self) -> &str { "extra-usage" }
-    fn description(&self) -> &str { "Show detailed usage statistics: calls, cache, tools" }
-    fn help(&self) -> &str {
-        "Usage: /extra-usage\n\n\
-         Displays extended usage statistics beyond /cost:\n\
-         - API call count\n\
-         - Cache hit/miss ratio\n\
-         - Token breakdown by type\n\
-         - Effective cost per call"
-    }
-
-    async fn execute(&self, _args: &str, ctx: &mut CommandContext) -> CommandResult {
-        let input = ctx.cost_tracker.input_tokens();
-        let output = ctx.cost_tracker.output_tokens();
-        let cache_creation = ctx.cost_tracker.cache_creation_tokens();
-        let cache_read = ctx.cost_tracker.cache_read_tokens();
-        let total = ctx.cost_tracker.total_tokens();
-        let cost = ctx.cost_tracker.total_cost_usd();
-
-        // Estimate API calls from messages (each assistant message ~ 1 API call)
-        let api_calls = ctx.messages.iter()
-            .filter(|m| m.role == claurst_core::types::Role::Assistant)
-            .count();
-        let api_calls = api_calls.max(1); // at least 1 if we have any data
-
-        // Cache efficiency
-        let cache_total = cache_creation + cache_read;
-        let cache_hit_pct = if cache_total > 0 {
-            (cache_read as f64 / cache_total as f64) * 100.0
-        } else {
-            0.0
-        };
-
-        let cost_per_call = if api_calls > 0 {
-            cost / api_calls as f64
-        } else {
-            0.0
-        };
-
-        CommandResult::Message(format!(
-            "Detailed Usage Statistics\n\
-             ─────────────────────────\n\
-             API calls:           {api_calls}\n\
-             Avg cost/call:       ${cost_per_call:.4}\n\n\
-             Token Breakdown:\n\
-               Input tokens:      {input:>10}\n\
-               Output tokens:     {output:>10}\n\
-               Cache creation:    {cache_creation:>10}\n\
-               Cache read:        {cache_read:>10}\n\
-               Total tokens:      {total:>10}\n\n\
-             Cache Performance:\n\
-               Cache hit rate:    {cache_hit_pct:.1}%\n\
-               Cache efficiency:  {cache_eff}\n\n\
-             Cost:\n\
-               Total cost:        ${cost:.4}\n\
-               Cost/1k tokens:    ${cost_per_k:.4}",
-            api_calls = api_calls,
-            cost_per_call = cost_per_call,
-            input = input,
-            output = output,
-            cache_creation = cache_creation,
-            cache_read = cache_read,
-            total = total,
-            cache_hit_pct = cache_hit_pct,
-            cache_eff = if cache_hit_pct > 70.0 {
-                "Excellent"
-            } else if cache_hit_pct > 40.0 {
-                "Good"
-            } else if cache_total > 0 {
-                "Low — prompts may not be stable enough to cache"
-            } else {
-                "No cache activity"
-            },
-            cost = cost,
-            cost_per_k = if total > 0 { cost / (total as f64 / 1000.0) } else { 0.0 },
-        ))
     }
 }
